requirements.txt
pycryptodome
peewee
python-decouple
peewee_migrate
